rename : 동행 package 구조 변경
