+++ src/main/java/com/crude/travelcrew/domain/crew/model/dto/CrewListRes.java
@@ -1,9 +1,9 @@
-package com.crude.travelcrew.domain.crew.dto;
+package com.crude.travelcrew.domain.crew.model.dto;
 
 import java.time.LocalDate;
 
-import com.crude.travelcrew.domain.crew.constants.CrewStatus;
-import com.crude.travelcrew.domain.crew.entity.Crew;
+import com.crude.travelcrew.domain.crew.model.constants.CrewStatus;
+import com.crude.travelcrew.domain.crew.model.entity.Crew;
 import com.crude.travelcrew.domain.member.entity.Member;
 
 import lombok.AllArgsConstructor;
