+++ src/main/java/com/crude/travelcrew/domain/crew/model/dto/CrewReq.java
@@ -1,10 +1,10 @@
-package com.crude.travelcrew.domain.crew.dto;
+package com.crude.travelcrew.domain.crew.model.dto;
 
 import java.time.LocalDate;
 
 
-import com.crude.travelcrew.domain.crew.constants.CrewGender;
-import com.crude.travelcrew.domain.crew.constants.CrewStatus;
+import com.crude.travelcrew.domain.crew.model.constants.CrewGender;
+import com.crude.travelcrew.domain.crew.model.constants.CrewStatus;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
