+++ src/main/java/com/crude/travelcrew/domain/crew/model/dto/MemberListReq.java
@@ -1,4 +1,4 @@
-package com.crude.travelcrew.domain.crew.dto;
+package com.crude.travelcrew.domain.crew.model.dto;
 
 public class MemberListReq {
 	private Long memberId;
