+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewCommentRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.crude.travelcrew.domain.crew.entity.CrewComment;
+import com.crude.travelcrew.domain.crew.model.entity.CrewComment;
 
 public interface CrewCommentRepository extends JpaRepository<CrewComment, Long> {
 	List<CrewComment> findByCrewId(long crewId, Pageable pageable);
