+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.crude.travelcrew.domain.crew.entity.Crew;
+import com.crude.travelcrew.domain.crew.model.entity.Crew;
 import com.crude.travelcrew.domain.member.entity.Member;
 
 public interface CrewRepository extends JpaRepository<Crew, Long> {
