+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewScrapRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.crude.travelcrew.domain.crew.entity.Crew;
-import com.crude.travelcrew.domain.crew.entity.CrewScrap;
+import com.crude.travelcrew.domain.crew.model.entity.Crew;
+import com.crude.travelcrew.domain.crew.model.entity.CrewScrap;
 import com.crude.travelcrew.domain.member.entity.Member;
 
 public interface CrewScrapRepository extends JpaRepository<CrewScrap, Long> {
