+++ src/main/java/com/crude/travelcrew/domain/crew/service/CrewService.java
@@ -11,13 +11,13 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.crude.travelcrew.domain.crew.dto.CrewCommentReq;
-import com.crude.travelcrew.domain.crew.dto.CrewCommentRes;
-import com.crude.travelcrew.domain.crew.dto.CrewListRes;
-import com.crude.travelcrew.domain.crew.dto.CrewReq;
-import com.crude.travelcrew.domain.crew.dto.CrewRes;
-import com.crude.travelcrew.domain.crew.entity.CrewComment;
-import com.crude.travelcrew.domain.crew.entity.Crew;
+import com.crude.travelcrew.domain.crew.model.dto.CrewCommentReq;
+import com.crude.travelcrew.domain.crew.model.dto.CrewCommentRes;
+import com.crude.travelcrew.domain.crew.model.dto.CrewListRes;
+import com.crude.travelcrew.domain.crew.model.dto.CrewReq;
+import com.crude.travelcrew.domain.crew.model.dto.CrewRes;
+import com.crude.travelcrew.domain.crew.model.entity.CrewComment;
+import com.crude.travelcrew.domain.crew.model.entity.Crew;
 import com.crude.travelcrew.domain.crew.repository.CrewCommentRepository;
 import com.crude.travelcrew.domain.crew.repository.CrewRepository;
 import com.crude.travelcrew.domain.member.entity.Member;
